+++ libs/designer/src/domains/ui/reducers/left-sidebar.ts
@@ -1,7 +1,8 @@
 import { DesignerEvent } from "@paperclip-ui/designer/src/events";
-import { DesignerState, selectNode } from "@paperclip-ui/designer/src/state";
+import { DesignerState } from "@paperclip-ui/designer/src/state";
 import { ast } from "@paperclip-ui/proto-ext/lib/ast/pc-utils";
 import produce from "immer";
+import { selectNode } from "../state";
 
 export const leftSidebarReducer = (
   state: DesignerState,
