+++ libs/designer/src/state/index.ts
@@ -1,2 +1,3 @@
 export * from "./core";
 export * from "./pc";
+export * from "./ui";
